+++ packages/calcite-components/src/components.d.ts
@@ -1969,6 +1969,10 @@ export namespace Components {
          */
         "beforeBack": () => Promise<void>;
         /**
+          * Passes a function to run before the component closes.
+         */
+        "beforeClose": () => Promise<void>;
+        /**
           * When `true`, displays a close button in the trailing side of the component's header.
          */
         "closable": boolean;
@@ -3767,6 +3771,10 @@ export namespace Components {
         "totalItems": number;
     }
     interface CalcitePanel {
+        /**
+          * Passes a function to run before the component closes.
+         */
+        "beforeClose": () => Promise<void>;
         /**
           * When `true`, displays a close button in the trailing side of the header.
          */
@@ -9948,6 +9956,10 @@ declare namespace LocalJSX {
          */
         "beforeBack"?: () => Promise<void>;
         /**
+          * Passes a function to run before the component closes.
+         */
+        "beforeClose"?: () => Promise<void>;
+        /**
           * When `true`, displays a close button in the trailing side of the component's header.
          */
         "closable"?: boolean;
@@ -11847,6 +11859,10 @@ declare namespace LocalJSX {
         "totalItems"?: number;
     }
     interface CalcitePanel {
+        /**
+          * Passes a function to run before the component closes.
+         */
+        "beforeClose"?: () => Promise<void>;
         /**
           * When `true`, displays a close button in the trailing side of the header.
          */