+++ packages/calcite-components/src/components/shell/shell.tsx
@@ -1,4 +1,4 @@
-import { Component, Element, Fragment, h, Listen, Prop, State, VNode } from "@stencil/core";
+import { Component, Element, Fragment, h, Listen, Prop, State, VNode, Watch } from "@stencil/core";
 import {
   ConditionalSlotComponent,
   connectConditionalSlotComponent,
@@ -77,7 +77,19 @@ export class Shell implements ConditionalSlotComponent {
 
   @State() hasSheets = false;
 
+  @State() hasPanelTop = false;
+
+  @State() hasPanelBottom = false;
+
+  @State() hasOnlyPanelBottom = false;
+
   @State() panelIsResizing = false;
+
+  @Watch("hasPanelTop")
+  @Watch("hasPanelBottom")
+  updateHasOnlyPanelBottom(): void {
+    this.hasOnlyPanelBottom = !this.hasPanelTop && this.hasPanelBottom;
+  }
 
   // --------------------------------------------------------------------------
   //
@@ -132,6 +144,14 @@ export class Shell implements ConditionalSlotComponent {
         (el as HTMLCalciteModalElement).slottedInShell = true;
       }
     });
+  };
+
+  handlePanelTopChange = (event: Event): void => {
+    this.hasPanelTop = slotChangeHasAssignedElement(event);
+  };
+
+  handlePanelBottomChange = (event: Event): void => {
+    this.hasPanelBottom = slotChangeHasAssignedElement(event);
   };
 
   handleDialogsSlotChange = (event: Event): void => {
@@ -208,8 +228,16 @@ export class Shell implements ConditionalSlotComponent {
     const deprecatedCenterRowSlotNode: VNode = (
       <slot key="center-row-slot" name={SLOTS.centerRow} />
     );
-    const panelBottomSlotNode: VNode = <slot key="panel-bottom-slot" name={SLOTS.panelBottom} />;
-    const panelTopSlotNode: VNode = <slot key="panel-top-slot" name={SLOTS.panelTop} />;
+    const panelBottomSlotNode: VNode = (
+      <slot
+        key="panel-bottom-slot"
+        name={SLOTS.panelBottom}
+        onSlotchange={this.handlePanelBottomChange}
+      />
+    );
+    const panelTopSlotNode: VNode = (
+      <slot key="panel-top-slot" name={SLOTS.panelTop} onSlotchange={this.handlePanelTopChange} />
+    );
 
     const contentContainerKey = "content-container";
 
@@ -224,14 +252,22 @@ export class Shell implements ConditionalSlotComponent {
           >
             {defaultSlotContainerNode}
           </div>,
-          <div class={CSS.contentBehindCenterContent}>
+          <div
+            class={{
+              [CSS.contentBehindCenterContent]: true,
+              [CSS.contentBottom]: this.hasOnlyPanelBottom,
+            }}
+          >
             {panelTopSlotNode}
             {panelBottomSlotNode}
             {deprecatedCenterRowSlotNode}
           </div>,
         ]
       : [
-          <div class={CSS.content} key={contentContainerKey}>
+          <div
+            class={{ [CSS.content]: true, [CSS.contentBottom]: this.hasOnlyPanelBottom }}
+            key={contentContainerKey}
+          >
             {panelTopSlotNode}
             {defaultSlotContainerNode}
             {panelBottomSlotNode}